Move Runtime syntax to syntax sub-folder
