+++ casper/src/main/scala/coop/rchain/casper/rholang/syntax/RuntimeReplaySyntax.scala
@@ -1,4 +1,4 @@
-package coop.rchain.casper.rholang
+package coop.rchain.casper.rholang.syntax
 
 import cats.data.EitherT
 import cats.effect.Sync
@@ -13,16 +13,16 @@ import coop.rchain.casper.protocol.{
   ProcessedSystemDeploy,
   SlashSystemDeployData
 }
-import coop.rchain.casper.rholang.RuntimeSyntax.SysEvalResult
+import coop.rchain.casper.rholang.syntax.RuntimeSyntax.SysEvalResult
 import coop.rchain.casper.util.EventConverter
 import coop.rchain.casper.util.rholang.InterpreterUtil.printDeployErrors
+import coop.rchain.casper.util.rholang._
 import coop.rchain.casper.util.rholang.costacc.{
   CloseBlockDeploy,
   PreChargeDeploy,
   RefundDeploy,
   SlashDeploy
 }
-import coop.rchain.casper.util.rholang._
 import coop.rchain.catscontrib.Catscontrib._
 import coop.rchain.metrics.implicits._
 import coop.rchain.metrics.{Metrics, Span}
